src/tools/embeddings.py
"""Vertex AI Text Embeddings client."""

from typing import List
import time
import random
import logging
from google.cloud import aiplatform
from vertexai.language_models import TextEmbeddingModel

logger = logging.getLogger(__name__)


def get_embeddings(
    texts: List[str],
    dim: int = 1536,
    model_name: str = "text-embedding-004"
) -> List[List[float]]:
    """Get embeddings for texts using Vertex AI.
    
    Args:
        texts: List of texts to embed
        dim: Output dimensionality (768, 1536, or 3072)
        model_name: Embedding model name
    
    Returns:
        List of embedding vectors
    """
    # TODO: Implement embedding generation with Vertex AI
    # Use output_dimensionality parameter
    # Batch size 1-8, retry on 429/5xx
    pass


def batch_embed_with_retry(
    model: TextEmbeddingModel,
    texts: List[str],
    batch_size: int = 8,
    max_retries: int = 3
) -> List[List[float]]:
    """Embed texts in batches with retry logic.
    
    Args:
        model: Embedding model instance
        texts: Texts to embed
        batch_size: Batch size for API calls
        max_retries: Maximum retry attempts
    
    Returns:
        List of embedding vectors
    """
    # TODO: Implement batching and retry logic
    pass