src/tools/repo_io.py
"""Repository I/O operations - clone, list files, read files."""

import os
import shutil
import tempfile
from pathlib import Path
from typing import List, Optional
import git
import logging

logger = logging.getLogger(__name__)


def clone_repo(url: str, ref: Optional[str] = None) -> str:
    """Clone a repository and return local path.
    
    Args:
        url: Repository URL
        ref: Branch, tag, or commit SHA
    
    Returns:
        Local path to cloned repository
    """
    # TODO: Implement repo cloning with GitPython
    pass


def list_files(
    local_path: str,
    include_globs: Optional[List[str]] = None,
    exclude_globs: Optional[List[str]] = None
) -> List[str]:
    """List files in repository matching patterns.
    
    Args:
        local_path: Path to repository
        include_globs: Patterns to include
        exclude_globs: Patterns to exclude
    
    Returns:
        List of relative file paths
    """
    # TODO: Implement file listing with glob patterns
    pass


def read_file(local_path: str, file_path: str) -> str:
    """Read file contents from repository.
    
    Args:
        local_path: Repository root path
        file_path: Relative path to file
    
    Returns:
        File contents as string
    """
    # TODO: Implement safe file reading
    pass