src/tools/retrieval.py
"""Hybrid retrieval with BM25 and vector search."""

from typing import List, Dict, Optional
import logging
from rank_bm25 import BM25Okapi
import numpy as np

logger = logging.getLogger(__name__)


class HybridRetriever:
    """Hybrid retrieval combining BM25 and vector search."""
    
    def __init__(self, embeddings_tool, vector_search_tool):
        self.embeddings = embeddings_tool
        self.vector_search = vector_search_tool
        self.bm25_index = None
        self.chunks = []
        
    def bm25_index_chunks(self, chunks: List[Dict]) -> None:
        """Build BM25 index from chunks.
        
        Args:
            chunks: List of chunk dictionaries
        """
        # TODO: Build BM25 index using rank_bm25
        pass
        
    def ann_index(self, vectors: List[List[float]]) -> None:
        """Build ANN index from vectors.
        
        Args:
            vectors: List of embedding vectors
        """
        # TODO: Build in-memory ANN index or use Vector Search
        pass
        
    def search(
        self,
        query_text: str,
        k: int = 12,
        expand_neighbors: bool = True
    ) -> List[Dict]:
        """Hybrid search with reciprocal rank fusion.
        
        Args:
            query_text: Search query
            k: Number of results
            expand_neighbors: Whether to include neighbor chunks
        
        Returns:
            Ranked retrieval results
        """
        # TODO: Implement hybrid search
        # 1. Embed query
        # 2. Get ANN candidates
        # 3. Get BM25 candidates
        # 4. Merge with reciprocal rank fusion
        # 5. Expand with neighbors from CodeMap
        pass
        
    def reciprocal_rank_fusion(
        self,
        ann_results: List[Dict],
        bm25_results: List[Dict],
        k: int = 60
    ) -> List[Dict]:
        """Merge results using reciprocal rank fusion.
        
        Args:
            ann_results: Results from vector search
            bm25_results: Results from BM25
            k: Fusion constant
        
        Returns:
            Merged and re-ranked results
        """
        # TODO: Implement RRF merging
        pass