pyproject.toml
[project]
name = "adk-multi-agent"
version = "0.1.0"
description = "Multi-agent system on GCP with Google ADK"
requires-python = ">=3.10"
dependencies = [
    "google-adk",  # ADK framework
    "google-cloud-aiplatform",
    "vertexai",
    "pydantic>=2.0",
    "httpx",
    "gitpython",
    "pyyaml",
    "redis",
    "uvicorn",
    "fastapi",
    "rank-bm25",
    "google-cloud-secret-manager",
    # "faiss-cpu",  # Optional, uncomment if needed
]

[project.optional-dependencies]
dev = [
    "ruff~=0.5",
    "mypy~=1.10",
    "pytest~=8.3",
    "pytest-asyncio~=0.23",
    "pytest-cov~=5.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.ruff]
line-length = 100
target-version = "py310"
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501", # line too long (handled by formatter)
    "B008", # do not perform function calls in argument defaults
]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --tb=short"